Support building by BUILD_TYPE where the default is rake but can optionally be the i18ntools generate script for localization template generation.

+++ buildtools/remote-dev-build.sh
@@ -12,6 +12,14 @@ echo "deb http://package-server/public/$REPOSITORY $DISTRIBUTION main non-free" 
 apt-get update
 apt -o Dpkg::Options::="--force-overwrite" build-dep -y --host-architecture $ARCHITECTURE .
 
-echo "RAKE_LOG=$RAKE_LOG"
-rm -f $RAKE_LOG
-rake |& tee $RAKE_LOG
+if [ "$BUILD_TYPE" = "rake" ] ; then
+  # Full uvm build
+  echo "RAKE_LOG=$RAKE_LOG"
+  rm -f $RAKE_LOG
+  rake |& tee $RAKE_LOG
+elif [ "$BUILD_TYPE" = "i18n" ] ; then
+  # i18n template build
+  cd i18ntools
+  ./generate.py
+  cd -
+fi
